Broken GLSL 100 shaders
There is something wrong with the mv_matrix in the GLSL 100 shader on the
raspberry pi. If I change the shader to use the mvp matrix then I get a
distorted (expected) lighting on the cube, but the mv matrix stops rendering
all-together. At a loss. This matrix works on the desktop version.

Perhaps there is a limit on the nr of matricies available as uniforms.