The nightmare of textures continues
OpenGL is horrible. Horrible. Pure horror. If I mess with G_TEXTURE[x] I can
get the screen fragment shader to accidentially sample from the BMP wall
textures, proving that the pixel formats are correct and the texture is being
correctly moved from client code to server. So this pins the issue to being
related to rendering the scene to a texture, and also using a texture in that
scene itself. All information online is either for different versions of GLSL,
or uses built-in variables like gl_TexCoord[0] which is almost certainly not
going to work in OpenGLES.