Skybox rendering
The skybox was messed up due to my indexes starting from 1 not 0 (Duh). This
commit also includes changes to which projection matrix is used for the skybox.
There is a third camera that represents a view matrix without any IOD offset.
This is supposed to mean that regardless of it's physical size, the skybox is
always rendered without any difference between the left and right eye views. So
it always looks like it is at infinity. This needs more work because these
projection matricies are being passed around too much.