Basic implementation of render virtual machine
This implements a very basic virtual machine for the scene renderer. Currently
vroom uses an array of 3 32 bit integers to represent the rendering pipeline -
the memory_id where the matrix is, the index offset into that memory for the
matrix and an object_id to render. This will be replaced with a virtual machine
that can execute opcodes for loading matrix data, accessing objects and in the
future opcodes enabling loops.