Major rewrite of the API around data handling
Data objects are now treated like arrays. They have an item_length which is the byte size of an item in that array, and a data_length which is the byte size of a single bit of data in that item. For example, a data object representing a vertex array of 4 verticies:

  memory_length:  48  (item_length * 4)
  item_length:    12  (data_length * 3)
  data_length:    4   (sizeof(float))

A program object also now has a data object backing it. For example a program of 10 bytecodes would look like:

  memory_length:  10  (item_length * 10)
  item_length:    1   (data_length * 1)
  data_length:    1   (sizeof(uint8_t))

An RGB8 texture of 256x256 size:

  memory_length:  195075  (item_length * (255 * 255))
  item_length:    3       (data_length * 3 bytes per pixel)
  data_length:    1       (sizeof(uint8_t))

In this way the contents of a data object can be verified against the object being created from it.