Add TODO comment to src/vrms_render_vm.c
After changing the matrix registers from an array of pointers to an array of 16
* sizeof(float) values, it is now required to copy matrix data from memory into
the registers. This kinda sucks because the copy will happen a lot. It will be
better for draw() to go back to an array of pointers to matricies however this
is bad for pure VM matrix operations because they will then need to operate on
pointers, not a static list (what if a register points to nothing?).

An option is to "mount" data object matricies in a separate register set and
create a new set of VM opcodes for dealing with mounted matricies, but I hate
to create new opcodes just for that. I just need to make a decision and stick
to it...