Test pre-calculated texture shift for Oculus
The horizontal center of the eye in each screen side is not the center of the
screen side. The horizontal center of the eye on the screen will be 1/2 the IOD
(InterOcular Distance) from the center of the physical screen in either
direction. The Oculus rift DK1 has a 7 inch screen, which is 1.7 dm
(decimeters). With an IOD of 0.66 dm the texture shift will be (1.7 / 4) - (IOD
/ 2) == 0.095 dm (almost 1 cm).