Remove the need for -rdynamic in main binary
The flag was required so modules could call back into the runtime to do things.
However it meant that every symbol in the server was callable from a module.
This change already caught one unintended thing which was the input_openhmd
module was accidentially calling into the a server esm function.

The new way creates an interface struct with a list of all allowed function
pointers a module can call.