Remove screen size hack
Previous way of detecting Oculus screen was that it reports a zero physical
width (!). Hopefully this works better as it will skip anything that uses an
"internal" connector. In the bright FUTURE vroom should be made to support
multiple screens and do sensible things when there is more than one. For
example a non-stero render of the scene.